organizing code and translating portuguese comments
