+++ src/components/Chart/adaptDomainToData.test.ts
